Use alternatives to TS "as"
In the case of the Content component it's being used to mark data as
defined when the status is "success", which is ensured. Using "!" makes
for a more terse code.

In the case of hooks, the URL is not used if login is undefined, but as
it's still a viable  value, it's cast into a string instead of using the
"!" type guard.